+++ app/javascript/components/widget/widget-templates/minimun-distance/config.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { createPortal } from 'react-dom';
 import Legend from 'components/widget/legend';
+import Tooltip from 'components/widget/tooltip';
 
 const sortData = data => data.sort((a, b) => a.value - b.value).reverse();
 
@@ -58,6 +59,19 @@ export const CONFIG = {
           content: () => {
             return createPortal(<Legend data={chartData} />, document.querySelector(`#widget-legend-${id}`));
           }
+        },
+        tooltip: {
+          cursor: false,
+          content: (
+            <Tooltip
+              style={{
+                flexDirection: 'column',
+                marginTop: '10px',
+                marginLeft: '-50px'
+              }}
+              payload={[chartData]}
+            />
+          )
         }
       }
     };