+++ app/javascript/components/widget/widget-templates/number-of-services/constants.js
@@ -1,5 +1,0 @@
-const defaultColors = {
-  'Mobile Network Operator': { color: 'rgb(214, 39, 40)' },
-  MFI: { color: 'rgb(44, 160, 44)' }
-};
-export default defaultColors;